Correciones

+++ crud_rest_db/src/main/java/edu/mondragon/we2/pinkAlert/controller/DoctorController.java
@@ -69,7 +69,7 @@ public class DoctorController {
 
         List<DatePill> datePills = IntStream.rangeClosed(0, 5)
                 .mapToObj(offset -> {
-                    LocalDate date = today.minusDays(5 - offset);
+                    LocalDate date = today.minusDays((long)(5 - offset));
 
                     long daysDiff = ChronoUnit.DAYS.between(date, today);
